pyproject.toml
[tool.poetry]
name = "starling-app"
version = "0.1.0"
packages = [{ include = "src" }]

description = ""
authors = ["ricardo.vaz <ricardodaniel.vaz@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
httpx = "^0.27.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-httpx = "^0.30.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
app = 'src.main:start'
app2 = 'src.main2:start'